+++ lib/myMethods.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+Text categories(String text) {
+  return Text(
+    '$text',
+    style: TextStyle(
+      color: Colors.red,
+    ),
+  );
+}
